Added images prop
