+++ components/bigcommerce/actions/delete-product/delete-product.mjs
@@ -5,7 +5,7 @@ export default {
   name: "Delete Product",
   description:
     "Delete a product by Id. [See the docs here](https://developer.bigcommerce.com/api-reference/76f5ebcdab695-delete-a-product)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   props: {
     bigcommerce,
