+++ components/bigcommerce/common/constants.mjs
@@ -190,7 +190,24 @@ const WEBHOOK_SCOPES = {
   },
 };
 
+const FILE_PROP_NAMES = [
+  "image_file",
+];
+
+const CONTENT_TYPE_KEY_HEADER = "Content-Type";
+const MULTIPART_FORM_DATA_VALUE_HEADER = "multipart/form-data";
+const MULTIPART_FORM_DATA_HEADERS = {
+  [CONTENT_TYPE_KEY_HEADER]: MULTIPART_FORM_DATA_VALUE_HEADER,
+};
+
+const DEFAULT_LIMIT = 100;
+
 export default {
   BASE_URL,
   WEBHOOK_SCOPES,
+  FILE_PROP_NAMES,
+  CONTENT_TYPE_KEY_HEADER,
+  MULTIPART_FORM_DATA_VALUE_HEADER,
+  MULTIPART_FORM_DATA_HEADERS,
+  DEFAULT_LIMIT,
 };
