+++ components/bigcommerce/sources/new-order/new-order.mjs
@@ -3,7 +3,7 @@ import common from "../common/base.mjs";
 export default {
   ...common,
   name: "New Order",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "source",
   key: "bigcommerce-new-order",
   description: "Emit new created order",
